requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.1
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
